create, update and delete task. Fixes soon

+++ src/main/java/com/rodrigoleon/todos/service/TaskService.java
@@ -1,5 +1,6 @@
 package com.rodrigoleon.todos.service;
 
+import com.rodrigoleon.todos.exception.task.TaskDoesNotExistException;
 import com.rodrigoleon.todos.model.Task;
 import com.rodrigoleon.todos.repository.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,12 +32,22 @@ public class TaskService {
         return taskRepository.findById(id);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws TaskDoesNotExistException {
+        Optional<Task> task = taskRepository.findById(id);
+        if (task.isEmpty()) throw new TaskDoesNotExistException();
+
         taskRepository.deleteById(id);
     }
 
-    public Task updateById(Long id, Task updatedTask) {
-        return taskRepository.save(updatedTask);
+    public Task updateById(Long id, Task updatedTask) throws TaskDoesNotExistException {
+        Optional<Task> task = taskRepository.findById(id);
+        if (task.isEmpty()) throw new TaskDoesNotExistException();
+
+        task.get().setTitle(updatedTask.getTitle());
+        task.get().setCompleted(updatedTask.isCompleted());
+
+        return taskRepository.save(task.get());
     }
 
 }
+
